requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-jazzmin==2.6.0
django-modeltranslation==0.18.10
django-phonenumber-field==7.1.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
phonenumbers==8.13.15
Pillow==9.5.0
pyTelegramBotAPI==4.12.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
